stream_to_bigquery-cf-event/requirements.txt
functions-framework
numpy
scikit-learn
torch
transformers
google-cloud-bigquery
google-cloud-dataflow
google-cloud-speech
google-cloud-aiplatform
google-cloud-texttospeech
google-cloud-storage
firebase-admin
pinecone